backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
langchain>=0.1.0
langgraph>=0.0.26
langchain-google-genai>=0.0.9
PyMuPDF>=1.23.0
python-docx>=1.1.0
python-dotenv>=1.0.0
chromadb>=0.4.15
langchain-community>=0.0.10
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client   